requirements.txt
# Flask
# gunicorn
# Werkzeug
dash
pandas
numpy
scipy
pyyaml
plotly
us
azure-storage-blob